added method pointer exercise to functional

+++ latex/day4/problems/listings/functional_functor_square.cpp
@@ -1,4 +1,4 @@
 class Square {
 public:
-	int operator() (int i) { return i*i; }
+	double operator() (double i) { return i*i; }
 };
